fix for cyml: nint to int, removed where

+++ Platforms/Stics_soil_temperature/tests/f90/test_soiltemp_component_time_loop.f90
@@ -74,7 +74,7 @@ program test_soiltemp
    call get_soil_depth(layer_thick, soil_depth)
 
    ! Number of elemental layers (1 cm)
-   soil_dim = nint(soil_depth)
+   soil_dim = int(soil_depth)
 
    ! In Stics the soil temperature profile and the previous day canopy temperature
    ! are initialized with the mean temperature of day 1
